feat: Added Author Controller
